fix(persona): revise whole Parakeet ONNX voice turns
